+++ Assets/Player/Script/IWeapon.cs
@@ -4,9 +4,9 @@ public interface IWeapon
 {
     void Fire();
     void MakeKinematic(bool kine=true);
-
     Material GetMaterial();
     void SetMaterial(Material material);
     void Reload();
-
+    void SetLock(bool enabled);
+    bool IsLocked();
 }